Changes on Model and PersistanceStorage

+++ src/main/java/org/eidos/kingchallenge/controller/LoginController.java
@@ -1,10 +1,12 @@
 package org.eidos.kingchallenge.controller;
 
+import java.util.concurrent.atomic.AtomicLong;
+
 /**
  * 
  * @author eidos71
  *
  */
 public interface LoginController {
-	public String loginService (Integer token) throws RuntimeException;
+	public String loginService (AtomicLong token) throws RuntimeException;
 }
